docker/athanor-rag/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
chromadb>=0.5
openai>=1.0
httpx>=0.27
google-cloud-storage>=2.14
google-cloud-secret-manager>=2.20
networkx>=3.2
pyyaml>=6.0